+++ src/prefect/utilities/notifications/__init__.py
@@ -1,0 +1,8 @@
+from prefect.utilities.notifications.notifications import callback_factory
+from prefect.utilities.notifications.notifications import slack_notifier
+from prefect.utilities.notifications.notifications import gmail_notifier
+
+try:
+    from prefect.utilities.notifications.jira_notification import jira_notifier
+except ImportError:
+    pass
